refactor AppBar to display list of export WIP
