+++ src/app/js/export/availableExports.js
@@ -1,0 +1,3 @@
+export default [
+    'csv',
+];
